Обновление базы ч.2

+++ ShopAPI/Models/BuyForShop.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public int IdProduct { get; set; }
-        public int Price { get; set; }
+        public float Price { get; set; }
         public int Count { get; set; }
-        public int Amount { get; set; }
+        public float Amount { get; set; }
 
         public Product? Product { get; set; }
     }
